My Relays: Change "Advertise Relay List" button style to 'important button' style
